flask/requirements.txt
RPi.GPIO
flask
gunicorn
